Better collision resolution

+++ src/dev/blijde_broers/misc/math/ExtendedMath.java
@@ -2,8 +2,8 @@ package dev.blijde_broers.misc.math;
 
 public class ExtendedMath {
 
-	public static float random(float min, float max) {
-		float out = (float) Math.random();
+	public static double random(double min, double max) {
+		double out = Math.random();
 		out = min + (out * (max - min));
 		return out;
 	}
