+++ src/dev/blijde_broers/misc/math/Point.java
@@ -2,9 +2,9 @@ package dev.blijde_broers.misc.math;
 
 public class Point {
 	
-	public float x, y;
+	public double x, y;
 
-	public Point(float x, float y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
